# srcertback/web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "api srcert.ru",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.srcert.ru/index.php/",
            "description": "Api server"
        }
    ],
    "paths": {
        "/agreement": {
            "get": {
                "tags": [
                    "Agreement"
                ],
                "summary": "list agreements",
                "description": "find agreement",
                "operationId": "app\\controllers\\AgreementController::actionIndex",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id to find",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "query to find",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agreement"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/agreement/create": {
            "post": {
                "tags": [
                    "Agreement"
                ],
                "summary": "create  agreement ",
                "description": "create new agreement",
                "operationId": "app\\controllers\\AgreementController::actionCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Agreement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Agreement"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/agreement/delete": {
            "delete": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Delete agreement",
                "description": "delete  agreement",
                "operationId": "app\\controllers\\AgreementController::actionDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Agreement id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/agreement/restore": {
            "get": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Restore agreement",
                "description": "restore  agreement",
                "operationId": "app\\controllers\\AgreementController::actionRestore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Agreement id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/agreement/update": {
            "post": {
                "tags": [
                    "Agreement"
                ],
                "summary": "update  agreement ",
                "description": "update agreement",
                "operationId": "app\\controllers\\AgreementController::actionUpdate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Agreement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Agreement"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/application": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "list applications",
                "description": "find application",
                "operationId": "app\\controllers\\ApplicationController::actionIndex",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id to find",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "query to find",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Application"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/application/create": {
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "create  application ",
                "description": "create new application",
                "operationId": "app\\controllers\\ApplicationController::actionCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/application/delete": {
            "delete": {
                "tags": [
                    "Application"
                ],
                "summary": "Delete application",
                "description": "delete  application",
                "operationId": "app\\controllers\\ApplicationController::actionDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Application id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/application/restore": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Restore application",
                "description": "restore  application",
                "operationId": "app\\controllers\\ApplicationController::actionRestore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Application id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/application/update": {
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "update  application ",
                "description": "update application",
                "operationId": "app\\controllers\\ApplicationController::actionUpdate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "list articless",
                "operationId": "app\\controllers\\ArticleController::actionIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Articles"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/articles/create": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "create new  articles ",
                "operationId": "app\\controllers\\ArticleController::actionCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Articles"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Articles"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
        "/articles/delete": {
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Deletes a articles",
                "operationId": "app\\controllers\\ArticleController::actionDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Articles id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "PHPSESSID"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    }
}